feat: moved to drizzel first approach for datascructures & created first migration

+++ shared/types/src/primitives.ts
@@ -1,4 +1,4 @@
-import { z } from 'zod';
+import { z } from "zod";
 
 export const identifierSchema = z.string().min(1);
 export const isoDateSchema = z.iso.date();
@@ -6,7 +6,7 @@ export const isoDateTimeSchema = z.iso.datetime({ offset: true });
 export const optionalTextSchema = z.string().optional();
 export const decimalRatingSchema = z.number().min(1).max(10);
 export const rirSchema = z.number().min(0).max(10);
-export const secondsSchema = z.number().int().nonnegative();
+export const secondsSchema = z.number().nonnegative();
 export const smallIntSchema = z.number().int().nonnegative();
 export const kilogramsSchema = z.number().nonnegative();
 export const metersSchema = z.number().nonnegative();
